requirements.txt
allure-robotframework~=2.13.1
robotframework-requests==0.9.2
robotframework==6.0.2
robotframework-browser==16.2.0
robotframework-faker~=5.0.0
robotframework-jsonlibrary>=0.5
robotframework-pabot~=2.15.0
pyyaml~=5.4